A lot of small fixes and tweaks here and there. Added a barebones seed for basic functionality. Added a bit of documentation.

+++ app/Http/Controllers/GroupsController.php
@@ -15,6 +15,11 @@ class GroupsController extends Controller
 
     public function index()
     {
+        if(auth()->user()->role_id == 2)
+        {
+            abort(403);
+        }
+
         $groups = Group::all();
 
         return view('groups/index', [
@@ -22,13 +27,18 @@ class GroupsController extends Controller
         ]);
     }
 
-    public function store()
+    public function store(Request $request)
     {
+        if($request->user()->role_id == 2)
+        {
+            abort(403);
+        }
+
         request()->validate([
             'name' => ['required', 'min:5'],
             'description' => 'required',
             'year' => ['required', 'min:4', 'max:4'],
-            'semester' => ['required', 'min:5']
+            'semester' => ['required', 'string']
         ]);
 
         $group = new Group();
@@ -40,16 +50,26 @@ class GroupsController extends Controller
 
         $group->save();
 
-        return redirect('/groups');
+        return redirect('/groups')->with('added', 'Luokka lisätty.');
     }
 
     public function create()
     {
+        if(auth()->user()->role_id == 2)
+        {
+            abort(403);
+        }
+
         return view('/groups/create');
     }
 
     public function show($id)
     {
+        if(auth()->user()->role_id == 2)
+        {
+            abort(403);
+        }
+
         $group = Group::findOrFail($id);
 
         return view('/groups/index', [
@@ -57,8 +77,13 @@ class GroupsController extends Controller
         ]);
     }
 
-    public function update($id)
+    public function update(Request $request, $id)
     {
+        if($request->user()->role_id == 2)
+        {
+            abort(403);
+        }
+
         $group = Group::findOrFail($id);
 
         request()->validate([
@@ -75,20 +100,33 @@ class GroupsController extends Controller
 
         $group->save();
 
-        return redirect('/groups');
+        return redirect('/groups')->with('edited', 'Luokka muokattu.');
     }
 
+    // Delete functionality disabled.
+    /*
     public function destroy($id)
     {
+        if(auth()->user()->role_id == 2)
+        {
+            abort(403);
+        }
+
         $group = Group::findOrFail($id);
 
         $group->delete();
 
         return redirect('/groups');
     }
+    */
 
     public function edit($id)
     {
+        if(auth()->user()->role_id == 2)
+        {
+            abort(403);
+        }
+
         $group = Group::findOrFail($id);
 
         return view('groups/edit', [
